+++ CHANGELOG.md
@@ -1,5 +1,14 @@
-**3.6**  
+**3.6.2**  
+Changes:
+ - removed TOOLBAR_ICON. Use TOOLBAR instead.
+ 
+News:
+ - added support for custom cursor Drawable.
+  
+**3.6.1**  
+Fixes only.
 
+**3.6**  
 Changes:
 - SearchView.TOOLBAR_ICON is now deprecated. Use TOOLBAR instead.
 - SearchView.setText is now deprecated. Use setTextInput instead.